Measure the duration of a cmd in ns and print it in prompt
A small Rust program was developed that prints the duration in a human
readable form. This was also an experiment in how to reduce the size of
the generated binary. Running bloaty (github.com/google/bloaty) on the
release build showed that most of the size stems from debug information.
This [article][1] showed me how to strip this information from the binary.
I previously tried to set `-C debuginfo=0` option in RUSTFLAGS but it
did not change the size of the binary.

The prompt now looks like this:

```sh
s2 alinz@kn-workstation ~ [2s 575ms 240us 683ns]: echo foo
foo
s2 alinz@kn-workstation ~ [975us 534ns]:
```

[1]: https://lifthrasiir.github.io/rustlog/why-is-a-rust-executable-large.html