Integrate shell-timer into rusty-prompt
This also fixes the return code issue by saving $? in an intermediate
local variable. Otherwise the return code of the previous command would
have been overwritten by that of the `date` call (to get the `end`
time of the duration).
Command durations are now available in bash as well.
This change also reduced prompt load time quite a bit because only
binary, except date, has to be run.