Fix autocompletion glitch
This removes the coloring of the second prompt line to fix terminal width
calculation in zsh and therefore prevent a glitch that occured on some
autocompletion items.
ZSH does not now about color escape characters when calculating the
terminal width, so they have to be specified explicitly [1]. To workaround
this issue I simply removed the coloring of the second prompt line.

PS: This also changes the prompt to `ƒ: ` which I think looks pretty
neat.

1: https://unix.stackexchange.com/a/90876