Remove percent suffix from battery state
The prefix needs to be escapd in a ZSH prompt (%%) but not in bash.
It is pretty hard to determine which shell is runnign since things like
"ZSH_NAME" is not set when the prompt generation is happening and
"SHELL" is not set in bash, but got inherited if bash was run inside a
zsh... what a mess.