Implement commit message prepare script in Python
The bash script contained a bug that when ammending a commit message
which already contained a jira ticket it was added another time.
I first tried to fix this using grep magic but it looked to complicated
and complex shell scripts are never clever hence I decided to replace it
by a small Python script.