WIP: Make git todo more reliable and minimal
The previous attempt had some flaws, e.g. when a file was not present in
of the given branches, then grep/rg failed.
Note that some additional escaping was required to add the command as
alias in the gitconfig.  Here's how the original command looks like:

$ git diff --no-prefix master... | grep -E '^\+(\+\+|.*TODO)' | sed
's/^\+*\s*//'

Note: this does not work as an alias, no matter which escaping I use.