Replace git todo alias with git-todo command
Custom git subcommands can be created pretty easily, they just have to
be named `git-<command>` and must be availabe in your PATH.

The previous alias had some downsides, e.g. grep failed if at least one
file from git ls-files was not part of the branch.  I tried to replace
it by another shell command but it showed that this got complicated
and for some reason I could not successfully escape the regex I used in
the git alias definition.  Anyways, this solution is more robust and
easy to maintain.