Add Microsoft Teams Desktop file
Its desktop app is utter gargabe on Linux, Screensharing will not work
in Wayland, it crashes regularly and so on.  Most of the functions just
work okay when teams is run in the browser.  But don't get me started,
Teams is just the worst communication software I know of, it is so bad
that I just cannot imagine how something like this can get shipped,
especially from a giant like Microsoft.  What makes me wonder even more
is that they know how to write Electron apps which they have proved with
Visual Studio Code.